requirements.txt
streamlit==1.28.1
ollama==0.1.7
langchain==0.0.350
langchain-community==0.0.10
pydantic==2.5.0
python-multipart==0.0.6
markdown==3.5.1
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
python-docx==1.1.0
